+++ src/entrypoint.c
@@ -14,6 +14,8 @@ void __wrap_abort(void) {
 
 // Rust functions
 SEXP wrap__cal_sheet_names(SEXP path);
+SEXP wrap__cal_sheets_metadata(SEXP path);
+SEXP wrap__cal_is_worksheet(SEXP path, SEXP sheet);
 SEXP wrap__cal_read_sheet(SEXP path, SEXP sheet);
 SEXP wrap__cal_read_sheet_df(SEXP path, SEXP sheet, SEXP col_names, SEXP skip, SEXP fill_merged);
 SEXP wrap__cal_sheet_dims(SEXP path, SEXP sheet);
@@ -21,6 +23,8 @@ SEXP wrap__cal_merge_regions(SEXP path, SEXP sheet);
 
 static const R_CallMethodDef CallEntries[] = {
     {"wrap__cal_sheet_names", (DL_FUNC) &wrap__cal_sheet_names, 1},
+    {"wrap__cal_sheets_metadata", (DL_FUNC) &wrap__cal_sheets_metadata, 1},
+    {"wrap__cal_is_worksheet", (DL_FUNC) &wrap__cal_is_worksheet, 2},
     {"wrap__cal_read_sheet", (DL_FUNC) &wrap__cal_read_sheet, 2},
     {"wrap__cal_read_sheet_df", (DL_FUNC) &wrap__cal_read_sheet_df, 5},
     {"wrap__cal_sheet_dims", (DL_FUNC) &wrap__cal_sheet_dims, 2},